Extract module test.on_yubikey.cli_piv.test_generate_cert_and_csr
